+++ lib/src/transition_type.dart
@@ -17,9 +17,12 @@ enum TransitionType {
   BOTTOM_CENTER_ROUNDER,
   CENTER_ROUNDER,
 }
+
 enum StripTransitionType {
   LEFT_TO_RIGHT,
   RIGHT_TO_LEFT,
   TOP_TO_BOTTOM,
   BOTTOM_TO_TOP
 }
+
+enum AnimatedOn { onTap, onHover }